updated cmake and type error

+++ src/BOM/historical_weather.c
@@ -65,7 +65,7 @@ static int8_t BOM_ParseWeather(ReqData_TypeDef *stream,
     MakeDirectory("datasets/bom");
     MakeDirectory("datasets/bom/historical");
 
-    int16_t full_filename_size = BOM_STATION_FILENAME_SIZE * 2;
+    size_t full_filename_size = (size_t)(BOM_STATION_FILENAME_SIZE * 2);
     char filename[full_filename_size];
     snprintf(filename, full_filename_size,
              "datasets/bom/historical/%s.csv",
